modified:   lib/LEDStrip/LEDInfo.cpp should be malloc(3*sizeof(int)) instead of malloc(3*sizeof(double))

+++ lib/LEDStrip/LEDInfo.cpp
@@ -3,7 +3,7 @@
 
 LEDInfo::LEDInfo(void){
 	step = (double*)malloc(3*sizeof(double));
-	deltas = (int*)malloc(3*sizeof(double));
+	deltas = (int*)malloc(3*sizeof(int));
 	carryOver = (double*)malloc(3*sizeof(double));
 
 	isActive = false;
